+++ src/linter.ts
@@ -24,7 +24,12 @@ export interface IWikitextModel extends Monaco.editor.ITextModel {
 	linter?: ITextModelLinter;
 }
 
-const storageKey = 'codemirror-mediawiki-addons';
+/**
+ * 获取CM6存储
+ * @param key 键名
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const getCmObject = (key: string): any => getObject(`codemirror-mediawiki-${key}`);
 
 /**
  * 获取代码检查器
@@ -46,12 +51,13 @@ const getLinter = (monaco: typeof Monaco, model: IWikitextModel, parserConfig: C
 				monaco.editor.setModelMarkers(model, 'Linter', diagnostics);
 				linter.glyphs = model.deltaDecorations(
 					linter.glyphs,
-					diagnostics.map(({startLineNumber, severity}) => ({
+					diagnostics.map(({startLineNumber, severity, message}) => ({
 						range: new monaco.Range(startLineNumber, 1, startLineNumber, 1),
 						options: {
 							glyphMarginClassName: `codicon codicon-${
 								severity === 8 as Monaco.MarkerSeverity ? 'error' : 'warning'
 							}`,
+							glyphMarginHoverMessage: {value: message},
 						},
 					})),
 				);
@@ -59,7 +65,7 @@ const getLinter = (monaco: typeof Monaco, model: IWikitextModel, parserConfig: C
 		}, 750);
 	};
 
-	if ((getObject(storageKey) as string[] | null)?.includes('lint')) {
+	if ((getCmObject('addons') as string[] | null)?.includes('lint')) {
 		const linter: ITextModelLinter = {glyphs: []};
 		(async () => {
 			switch (model.getLanguageId()) {
@@ -80,7 +86,12 @@ const getLinter = (monaco: typeof Monaco, model: IWikitextModel, parserConfig: C
 					break;
 				}
 				case 'javascript': {
-					const esLint: (text: string) => Linter.LintMessage[] = await getJsLinter();
+					const opt = {
+							env: {browser: true, es2024: true, jquery: true},
+							globals: {mw: 'readonly', mediaWiki: 'readonly', OO: 'readonly'},
+							...getCmObject('ESLint'),
+						} as Linter.Config as Record<string, unknown>,
+						esLint: (text: string) => Linter.LintMessage[] = await getJsLinter(opt);
 					linter.lint = (text): Monaco.editor.IMarkerData[] =>
 						esLint(text).map(({ruleId, message, severity, line, column, endLine, endColumn}) => ({
 							source: `ESLint(${ruleId})`,
@@ -94,7 +105,8 @@ const getLinter = (monaco: typeof Monaco, model: IWikitextModel, parserConfig: C
 					break;
 				}
 				case 'css': {
-					const styleLint: (code: string) => Promise<Warning[]> = await getCssLinter();
+					const opt: Record<string, unknown> | null = getCmObject('Stylelint'),
+						styleLint: (code: string) => Promise<Warning[]> = await getCssLinter(opt);
 					linter.lint = async (code): Promise<Monaco.editor.IMarkerData[]> =>
 						(await styleLint(code)).map(({text, severity, line, column, endLine, endColumn}) => ({
 							source: 'Stylelint',